ASNIDE-21 Moved files related to autocompletion to common directory
