+++ autocompletion/asnsnippetprovider.h
