+++ autocompletion/builtinsproposalsprovider.cpp
