+++ autocompletion/completionassist.cpp
@@ -33,15 +33,15 @@
 #include <texteditor/codeassist/genericproposal.h>
 #include <texteditor/codeassist/genericproposalmodel.h>
 
-#include "asn1acnconstants.h"
-#include "parseddatastorage.h"
-#include "parseddocument.h"
-
 #include "asnbuiltinsproposalsprovider.h"
 #include "acnbuiltinsproposalsprovider.h"
 
-#include "acneditor.h"
-#include "asneditor.h"
+#include "../asn1acnconstants.h"
+#include "../parseddatastorage.h"
+#include "../parseddocument.h"
+
+#include "../acneditor.h"
+#include "../asneditor.h"
 
 using namespace Asn1Acn::Internal;
