+++ autocompletion/completionassist.h
