+++ autocompletion/proposalsprovider.cpp
