+++ autocompletion/usertypesproposalsprovider.h
@@ -23,33 +23,34 @@
 **
 ****************************************************************************/
 
-#include "acncompletionassist.h"
+#pragma once
 
-#include "asn1acnconstants.h"
+#include <memory>
 
-#include "acnbuiltinsproposalsprovider.h"
+#include <texteditor/codeassist/assistproposalitem.h>
 
-using namespace Asn1Acn::Internal;
+#include "proposalsprovider.h"
+#include "../data/modules.h"
 
-AcnCompletionAssistProcessor::AcnCompletionAssistProcessor()
-    : CompletionAssistProcessor(QLatin1String(Constants::ACN1_SNIPPETS_GROUP_ID))
+namespace Asn1Acn {
+namespace Internal {
+
+class UserTypesProposalsProvider : public ProposalsProvider
 {
-}
+public:
+    UserTypesProposalsProvider(const std::unique_ptr<Data::Modules> &data);
 
-std::unique_ptr<BuiltinsProposalsProvider> AcnCompletionAssistProcessor::getBuiltinsProposalsProvider() const
-{
-    auto provider = std::make_unique<AcnBuiltinsProposalProvider>();
-    return std::move(provider);
-}
+private:
+    QList<TextEditor::AssistProposalItemInterface *> createProposals() const override;
 
-bool AcnCompletionAssistProvider::supportsEditor(Core::Id editorId) const
-{
-    return editorId == Constants::ACNEDITOR_ID;
-}
+    QList<TextEditor::AssistProposalItemInterface *>
+    createImportedTypes(const QList<QString> &importedProposals) const;
 
-TextEditor::IAssistProcessor *AcnCompletionAssistProvider::createProcessor() const
-{
-    return new AcnCompletionAssistProcessor;
-}
+    QList<TextEditor::AssistProposalItemInterface *>
+    createInternalTypes(const Data::Definitions::Types &types) const;
 
+    const std::unique_ptr<Data::Modules> &m_data;
+};
 
+} /* namespace Internal */
+} /* namespace Asn1Acn */
